move tests to dedicated dir
