+++ package.json
@@ -25,6 +25,10 @@
     "test:watch": "TZ=UTC vitest"
   },
   "author": "Defra DDTS",
+  "contributors": [
+    "John Watson <john.watson1@defra.gov.uk>",
+    "Rana Salem <rana.salem@defra.gov.uk>"
+  ],
   "license": "OGL-UK-3.0",
   "dependencies": {
     "@defra/hapi-tracing": "1.22.0",
