+++ test/unit/common/helpers/fail-action.test.js
@@ -1,4 +1,4 @@
-import { failAction } from './fail-action.js'
+import { failAction } from '../../../../src/common/helpers/fail-action.js'
 
 describe('#fail-action', () => {
   test('Should throw expected error', () => {
