+++ test/unit/common/helpers/metrics.test.js
@@ -1,7 +1,7 @@
 import { StorageResolution, Unit } from 'aws-embedded-metrics'
 
-import { config } from '../../config.js'
-import { metricsCounter } from './metrics.js'
+import { config } from '../../../../src/config.js'
+import { metricsCounter } from '../../../../src/common/helpers/metrics.js'
 
 const mockPutMetric = vi.fn()
 const mockFlush = vi.fn()
@@ -18,7 +18,7 @@ vi.mock('aws-embedded-metrics', async (importOriginal) => {
     })
   }
 })
-vi.mock('./logging/logger.js', () => ({
+vi.mock('../../../../src/common/helpers/logging/logger.js', () => ({
   createLogger: () => ({ error: (...args) => mockLoggerError(...args) })
 }))
