+++ test/unit/common/helpers/proxy/setup-proxy.test.js
@@ -1,6 +1,6 @@
-import { config } from '../../../config.js'
+import { config } from '../../../../../src/config.js'
 import { getGlobalDispatcher, ProxyAgent } from 'undici'
-import { setupProxy } from './setup-proxy.js'
+import { setupProxy } from '../../../../../src/common/helpers/proxy/setup-proxy.js'
 
 describe('setupProxy', () => {
   afterEach(() => {
