+++ test/unit/common/helpers/secure-context/get-trust-store-certs.test.js
@@ -1,4 +1,4 @@
-import { getTrustStoreCerts } from './get-trust-store-certs.js'
+import { getTrustStoreCerts } from '../../../../../src/common/helpers/secure-context/get-trust-store-certs.js'
 
 describe('#getTrustStoreCerts', () => {
   const mockProcessEnvWithCerts = {
