+++ test/unit/common/helpers/secure-context/secure-context.test.js
@@ -1,8 +1,8 @@
 import hapi from '@hapi/hapi'
 
-import { secureContext } from './index.js'
-import { requestLogger } from '../logging/request-logger.js'
-import { config } from '../../../config.js'
+import { secureContext } from '../../../../../src/common/helpers/secure-context/index.js'
+import { requestLogger } from '../../../../../src/common/helpers/logging/request-logger.js'
+import { config } from '../../../../../src/config.js'
 
 const mockAddCACert = vi.fn()
 const mockCreateSecureContext = vi
