+++ vitest.config.js
@@ -5,13 +5,23 @@ export default defineConfig({
     globals: true,
     environment: 'node',
     clearMocks: true,
+    include: ['**/test/**/*.test.js'],
     coverage: {
       provider: 'v8',
       reportsDirectory: './coverage',
       reporter: ['text', 'lcov'],
-      include: ['src/**'],
-      exclude: [...configDefaults.exclude, 'coverage']
-    },
-    setupFiles: ['.vite/setup-files.js']
+      include: ['src/**/*.js'],
+      exclude: [
+        ...configDefaults.exclude,
+        '.public',
+        'coverage',
+        'postcss.config.js',
+        'stylelint.config.js',
+        '**/node_modules/**',
+        '**/test/**',
+        '.server',
+        'src/index.js'
+      ]
+    }
   }
 })
